% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cHull.R
\name{cHull}
\alias{cHull}
\title{Convex Hull}
\usage{
cHull(x, y)
}
\arguments{
\item{x}{x of the coordinate pair}

\item{y}{y of the coordinate pair}
}
\value{
This function returns the convex hull area for the set of points provided.
}
\description{
This function returns the convex hull area for the set of points provided.
}
