% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eucDist.R
\name{eucDist}
\alias{eucDist}
\title{Euclidean Distance}
\usage{
eucDist(x1, y1, x2, y2)
}
\arguments{
\item{x1}{x of the first coordinate pair}

\item{y1}{y of the first coordinate pair}

\item{x2}{x of the second coordinate pair}

\item{y2}{y of the second coordinate pair}
}
\value{
This function returns a data frame containing a column of the calculated Euclidean distance
}
\description{
This function returns a data frame containing a column of the calculated Euclidean distance.
}
