% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formantAnalysis.R
\name{formantAnalysis}
\alias{formantAnalysis}
\title{Formant Analysis}
\usage{
formantAnalysis(time, formant)
}
\arguments{
\item{formant}{}
}
\value{

}
\description{
This function analyzes formant data.
}
